requirements.txt
kafka-python
langchain
langchain-community
neo4j
redis
dotenv
feedparser